+++ erc20-migration/hardhat.config.js
@@ -16,6 +16,9 @@ else {
     mnemonic: (process.env.MNEMONIC || "")
   }
 }
+
+let priorityFeeInWei = process.env.PRIORITY_FEE || "100"
+
 module.exports = {
   solidity: {
     version: "0.8.27",
@@ -172,8 +175,12 @@ task("contractSetup", "Deploys ZEN migration and vesting contracts").addFlag("ve
 
   const admin = (await ethers.getSigners())[0];
 
+  const overrides = {
+    maxPriorityFeePerGas: ethers.parseUnits(priorityFeeInWei, "wei")
+  };
+
   let factory = await hre.ethers.getContractFactory(ZEN_FACTORY_CONTRACT_NAME);
-  let ZenMigrationFactory = await factory.deploy(admin);
+  let ZenMigrationFactory = await factory.deploy(admin, overrides);
   let receipt = await ZenMigrationFactory.deploymentTransaction().wait(); // Wait for confirmation
 
   if (receipt.status == 0) {
@@ -186,6 +193,7 @@ task("contractSetup", "Deploys ZEN migration and vesting contracts").addFlag("ve
   let tokenName = process.env.TOKEN_NAME || "ZEN"
   let tokenSymbol = process.env.TOKEN_SYMBOL || "ZEN"
   let base_message = process.env.BASE_MESSAGE || "CLAIM"
+
   let res = await ZenMigrationFactory.deployMigrationContracts(
     tokenName,
     tokenSymbol,
@@ -193,7 +201,8 @@ task("contractSetup", "Deploys ZEN migration and vesting contracts").addFlag("ve
     process.env.HORIZEN_FOUNDATION_ADMIN,
     process.env.HORIZEN_FOUNDATION,
     process.env.HORIZEN_DAO_ADMIN,
-    process.env.HORIZEN_DAO
+    process.env.HORIZEN_DAO,
+    overrides
   );
 
   receipt = await res.wait();
@@ -403,8 +412,13 @@ task("restoreEON", "Restores EON accounts", async (taskArgs, hre) => {
 
   const EONVault = await hre.ethers.getContractAt(EON_VAULT_CONTRACT_NAME, process.env.EON_VAULT_ADDRESS);
 
+  const overrides = {
+    maxPriorityFeePerGas: ethers.parseUnits(priorityFeeInWei, "wei")
+  };
+
+
   console.log("Setting final account hash on EONVault");
-  let res = await EONVault.setCumulativeHashCheckpoint(finalCumAccountHash);
+  let res = await EONVault.setCumulativeHashCheckpoint(finalCumAccountHash, overrides);
   let receipt = await res.wait();
   if (receipt.status == 0) {
     console.error("Setting final account hash on EONVault failed! Failed transaction: " + res);
@@ -418,7 +432,7 @@ task("restoreEON", "Restores EON accounts", async (taskArgs, hre) => {
   console.log("                      Start loading accounts");
   console.log("***************************************************************\n\n");
 
-  const BATCH_LENGTH = 500;
+  const BATCH_LENGTH = 250;
   let addressesValues = [];
   let calcCumulativeHash = "0x0000000000000000000000000000000000000000000000000000000000000000";
   let batchNumber = 1;
@@ -434,7 +448,7 @@ task("restoreEON", "Restores EON accounts", async (taskArgs, hre) => {
     if (addressesValues.length == BATCH_LENGTH) {
       console.log(`Inserting batch ${batchNumber} of ${totalBatchNumber}`);
       try {
-        let res = await EONVault.batchInsert(calcCumulativeHash, addressesValues);
+        let res = await EONVault.batchInsert(calcCumulativeHash, addressesValues, overrides);
         let receipt = await res.wait();
         totalUsedGas = totalUsedGas + BigInt(receipt.gasUsed);
         console.log("Gas used: " + receipt.gasUsed);
@@ -460,7 +474,7 @@ task("restoreEON", "Restores EON accounts", async (taskArgs, hre) => {
   if (addressesValues.length > 0) {
     console.log(`Inserting last batch ${batchNumber} of ${totalBatchNumber}`);
     try {
-      let res = await EONVault.batchInsert(calcCumulativeHash, addressesValues);
+      let res = await EONVault.batchInsert(calcCumulativeHash, addressesValues, overrides);
       let receipt = await res.wait();
       totalUsedGas = totalUsedGas + BigInt(receipt.gasUsed);
       console.log("Gas used: " + receipt.gasUsed);
@@ -497,7 +511,7 @@ task("restoreEON", "Restores EON accounts", async (taskArgs, hre) => {
   let round = 0;
   while (await EONVault.moreToDistribute()) {
     console.log("Distribution round: " + round);
-    let res = await EONVault.distribute(process.env.DISTRIBUTE_MAX_COUNT);
+    let res = await EONVault.distribute(process.env.DISTRIBUTE_MAX_COUNT, overrides);
     let receipt = await res.wait();
     totalUsedGas = totalUsedGas + BigInt(receipt.gasUsed);
     console.log("Gas used: " + receipt.gasUsed);
@@ -580,8 +594,13 @@ task("restoreZEND", "Restores ZEND accounts", async (taskArgs, hre) => {
 
   const ZENDVault = await hre.ethers.getContractAt(ZEND_VAULT_CONTRACT_NAME, process.env.ZEND_VAULT_ADDRESS);
 
+
+  const overrides = {
+    maxPriorityFeePerGas: ethers.parseUnits(priorityFeeInWei, "wei")
+  };
+
   console.log("Setting final account hash on ZENDVault");
-  let res = await ZENDVault.setCumulativeHashCheckpoint(finalCumAccountHash);
+  let res = await ZENDVault.setCumulativeHashCheckpoint(finalCumAccountHash, overrides);
   let receipt = await res.wait();
   if (receipt.status == 0) {
     console.error("Setting final account hash on ZENDVault failed! Failed transaction: " + res);
@@ -610,7 +629,7 @@ task("restoreZEND", "Restores ZEND accounts", async (taskArgs, hre) => {
     if (addressesValues.length == BATCH_LENGTH) {
       console.log(`Inserting batch ${batchNumber} of ${totalBatchNumber}`);
       try {
-        let res = await ZENDVault.batchInsert(calcCumulativeHash, addressesValues);
+        let res = await ZENDVault.batchInsert(calcCumulativeHash, addressesValues, overrides);
         let receipt = await res.wait();
         totalUsedGas = totalUsedGas + BigInt(receipt.gasUsed);
         console.log("Gas used: " + receipt.gasUsed);
@@ -636,7 +655,7 @@ task("restoreZEND", "Restores ZEND accounts", async (taskArgs, hre) => {
   if (addressesValues.length > 0) {
     console.log(`Inserting last batch ${batchNumber} of ${totalBatchNumber}`);
     try {
-      let res = await ZENDVault.batchInsert(calcCumulativeHash, addressesValues);
+      let res = await ZENDVault.batchInsert(calcCumulativeHash, addressesValues, overrides);
       let receipt = await res.wait();
       totalUsedGas = totalUsedGas + BigInt(receipt.gasUsed);
       console.log("Gas used: " + receipt.gasUsed);