+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.3
+* Added (optional) priority fee override in migration scripts
+* Fixed contract verification in migration scripts
+
 ## 1.3.2
 * Added additional checks for signature format in VerificationLibrary
 * Added Testnet snapshot data and signatures